+++ components/embl-content-hub-loader/embl-content-hub-loader.js
@@ -12,6 +12,4 @@ function emblContentHub() {
   emblContentHubFetch();
 }
 
-emblContentHub();
-
 export { emblContentHub };