+++ fractal.js
@@ -1,8 +1,8 @@
 'use strict';
 
 /* Create a new Fractal instance and export it for use elsewhere if required */
-const fractal = module.exports = require('@frctl/fractal').create();
-const projectTitle = vfName;
+const fractal        = module.exports = require('@frctl/fractal').create();
+const projectTitle   = vfName;
 
 /* Set the title of the project */
 fractal.set('project.title', projectTitle);
@@ -13,51 +13,41 @@ fractal.components.set('path', __dirname + '/components');
 /* Tell Fractal where the documentation pages will live */
 fractal.docs.set('path', __dirname + '/docs');
 
-/* Handlebars with custom helpers */
-const handlebars = require('gulp-compile-handlebars');
-const hljs = require('highlight.js');
-const hbs = require('@frctl/handlebars')({
-  helpers: {
-    striptags: function(txt,context){
-      txt = txt.fn(context);
-      if(typeof txt == "undefined") return;
-      // the regular expresion
-      var regexp = /<[\/\w]+>/g
-      // replacing the text
-      return txt.replace(regexp, '');
-    },
-    escapetags: function(txt,context){
-      txt = txt.fn(context);
-      if(typeof txt == "undefined") return;
-      return handlebars.Handlebars.Utils.escapeExpression(txt);
-    },
-    codeblockhtml: function(txt,context){
-      txt = txt.fn(context);
-      if(typeof txt == "undefined") return;
-      return '<code class="Code Code--lang-html vf-code-example"><pre class="vf-code-example__pre">' +
-      hljs.highlight('html', txt).value + '</pre></code>';
-    },
-    codeblockjs: function(txt,context){
-      txt = txt.fn(context);
-      if(typeof txt == "undefined") return;
-      return '<code class="Code Code--lang-js vf-code-example"><pre class="vf-code-example__pre">' +
-      hljs.highlight('js', txt).value + '</pre></code>';
+const nunj = require('@frctl/nunjucks')({
+  env: {
+    lstripBlocks: true,
+    trimBlocks: true,
+    autoescape: false
+    // Nunjucks environment opts: https://mozilla.github.io/nunjucks/api.html#configure
+  },
+  filters: {
+    // A filter and non-async version of frctl's context extension from
+    // https://github.com/frctl/nunjucks/blob/develop/src/extensions/context.js
+    // We mainly use this to make a component's YAML data available to REAMDE.md
+    // {% set context = '@vf-heading' | componentContexts %}
+    componentContexts:  function(component) {
+      const source = fractal.components;
+      const handle = component;
+      const entity = source.find(handle);
+      if (!entity) {
+        throw new Error(`Could not render component '${handle}' - component not found.`);
+      }
+      const context = entity.isComponent ? entity.variants().default().context : entity.context;
+      return context;
     }
-    // bold: function(options) {
-    //   return new handlebars.Handlebars.SafeString(
-    //       '<strong class="mybold">'
-    //       + options.fn(this)
-    //       + '</strong>');
-    // },
-    // uppercase: function(str) {
-    //     return str.toUpperCase();
-    // }
+  },
+  // globals: {
+  //   // global-name: global-val
+  // },
+  extensions: {
+    codeblock: require('./tools/vf-frctl-extensions/codeblock.js')
   }
 });
 
-fractal.components.engine(hbs); /* set as the default template engine for components */
-fractal.docs.engine(hbs); /* you can also use the same instance for documentation, if you like! */
-
+fractal.components.set('ext', '.njk'); // look for files with a .nunj file extension
+fractal.components.engine(nunj); /* set as the default template engine for components */
+fractal.docs.set('ext', '.njk'); // look for files with a .njk file extension
+fractal.docs.engine(nunj); /* you can also use the same instance for documentation, if you like! */
 
 /* configure components */
 fractal.components.set('default.status', 'alpha');
@@ -74,17 +64,9 @@ fractal.web.set('server.syncOptions', {
   browser: 'default',
   sync: true
 });
-/* Theme */
-const mandelbrot = require('@frctl/mandelbrot');
 
-const VFTheme = mandelbrot({
-  // favicon: 'https://dev.assets.emblstatic.net/vf/assets/vf-favicon/assets/favicon.ico',
-  styles: [
-    '/local.css'
-  ],
-  format: 'YAML',
-  panels: ["html", "info", "resources"]
-});
+const vfTheme = require('./tools/vf-frctl-theme');
+const vfThemeConfig = vfTheme({}, fractal);
 
 fractal.components.set('statuses', {
   /* status definitions here */
@@ -111,17 +93,4 @@ fractal.components.set('statuses', {
   }
 });
 
-// Customise Fractal templates
-// https://fractal.build/guide/customisation/web-themes#template-customisation
-VFTheme.addLoadPath(__dirname + '/tools/frctl-mandelbrot-vf-subtheme/views');
-// Specify the static assets directory that contains the custom stylesheet.
-VFTheme.addStatic(__dirname + '/tools/frctl-mandelbrot-vf-subtheme/assets', '/');
-
-fractal.web.theme(VFTheme);
-
-// fractal.components.set('resources', {
-//   scss: {
-//     label: 'SCSS',
-//     match: ['**/*.scss']
-//   }
-// });
+fractal.web.theme(vfThemeConfig);