+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # CHANGELOG
 
-See https://visual-framework.github.io/vf-core/changelog
+See https://github.com/visual-framework/vf-core/releases