+++ components/embl-breadcrumbs-lookup/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.7 (2019-03-28)
+
+
+### Bug Fixes
+
+* don't insert a root term as a crumb ([003d88c](https://github.com/visual-framework/vf-core/commit/003d88c))
+* only insert crumb if it respects the original term context of who/what/where ([2159778](https://github.com/visual-framework/vf-core/commit/2159778))
+
+
+
+
+
 ## 0.0.6 (2019-02-28)
 
 **Note:** Version bump only for package @visual-framework/embl-breadcrumbs-lookup