+++ components/vf-font-plex-sans/CHANGELOG.md
@@ -3,9 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## 0.0.24 (2019-02-28)
+## 0.0.25 (2019-03-28)
 
 **Note:** Version bump only for package @visual-framework/vf-plex-sans-font
+
+
+
+
+
+## 0.0.24 (2019-02-28)
+
+**Note:** Version bump only for package @visual-framework/vf-font-plex-sans
 
 
 
@@ -13,7 +21,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## 0.0.23 (2019-02-28)
 
-**Note:** Version bump only for package @visual-framework/vf-plex-sans-font
+**Note:** Version bump only for package @visual-framework/vf-font-plex-sans